requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
NorenRestApi @ file:///home/pg/python/api_shoonya/dist/NorenRestApi-0.0.30-py2.py3-none-any.whl
numpy==2.2.6
pandas==2.3.3
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
